\name{StanTest-package}
\alias{StanTest-package}
\alias{StanTest}
\docType{package}
\title{
\packageTitle{StanTest}
}
\description{
\packageDescription{StanTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{StanTest}
\packageIndices{StanTest}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{StanTest}

Maintainer: \packageMaintainer{StanTest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
